Authentication and userendpoints

+++ src/TaskTracker.Api/TaskItemController.cs
@@ -3,10 +3,13 @@ using Microsoft.AspNetCore.Mvc;
 using TaskTracker.Domain;
 using Microsoft.EntityFrameworkCore;
 using TaskTracker.Contracts;
+using System.Security.Claims;
+using Microsoft.AspNetCore.Authorization;
 namespace TaskTracker.Api;
 
 [ApiController]
 [Route("api/[controller]")]
+[Authorize]
 
 public class TaskItemController : ControllerBase
 {
@@ -17,42 +20,42 @@ public class TaskItemController : ControllerBase
         _context = context;
     }
 
-    // Get
 
     [HttpGet]
-
     public async Task<ActionResult<IEnumerable<TaskItem>>> GetAllTaskAsync()
     {
-        return await _context.Tasks.ToListAsync();
+        var userId = User.FindFirstValue(ClaimTypes.NameIdentifier);
+        if(userId is null) return Unauthorized();
+        var tasks = _context.Tasks.Where(t => t.UserId == userId);
+        return Ok(tasks);
     }
 
     [HttpGet("{id}")]
     public async Task<ActionResult<TaskItem>> GetByIdAsync(int id)
     {
+        var userId = User.FindFirstValue(ClaimTypes.NameIdentifier);
+        if(userId is null) return Unauthorized();
         var task = await _context.Tasks.FindAsync(id);
         if(task is null) return NotFound();
+        if(task.UserId != userId) return Forbid();
+        
 
         return Ok(task);
     }
 
-    /*
     [HttpPost]
     public async Task<ActionResult<TaskItem>> CreateTaskAsync(CreateTaskItemRequest request)
     {
-        var rand = new Random();
-        
-        
-        
+        var userId = User.FindFirstValue(ClaimTypes.NameIdentifier);
+        if(userId is null) return Unauthorized();
+
         var taskItem = new TaskItem
         {
            CreatedAt = DateTimeOffset.UtcNow,
            Title = request.Title,
            Description = request.Description,
            Deadline = request.Deadline,
-           //  fix when register is up
-           UserId =, 
-        
-
+           UserId = userId 
         };
 
         _context.Add(taskItem);
@@ -61,13 +64,17 @@ public class TaskItemController : ControllerBase
         return CreatedAtAction(nameof(GetByIdAsync), new { id = taskItem.Id }, taskItem);
             
     }
-    */
-    
+   
     [HttpPut("{id}")]
     public async Task<IActionResult> UpdateTaskItem(int id, UpdateTaskItemRequest request)
     {
+        var userId = User.FindFirstValue(ClaimTypes.NameIdentifier);
+        if(userId is null) return Unauthorized();
+        
         var taskItem = await _context.Tasks.FindAsync(id);
         if(taskItem is null) return NotFound();
+
+        if(taskItem.UserId != userId) return Forbid();
 
         taskItem.Description = request.Description;
         taskItem.Completed = request.Completed;
@@ -78,12 +85,17 @@ public class TaskItemController : ControllerBase
         await _context.SaveChangesAsync();
         return NoContent();
     }
+
     [HttpDelete("{id}")]
     public async Task<IActionResult> DeleteTaskItem(int id)
     {
+        var userId = User.FindFirstValue(ClaimTypes.NameIdentifier);
+        if(userId is null) return Unauthorized();
+
         var taskItem = await _context.Tasks.FindAsync(id);
         if(taskItem is null) return NotFound();
 
+        if(taskItem.UserId != userId) return Forbid();
         _context.Tasks.Remove(taskItem);
         await _context.SaveChangesAsync();
